Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-canvas-tainting-video.https-expected.txt


PASS initialize global state
PASS url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&reject" with crossOrigin "" should be LOAD_ERROR
PASS url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&reject" with crossOrigin "anonymous" should be LOAD_ERROR
PASS url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&reject" with crossOrigin "use-credentials" should be LOAD_ERROR
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&ignore" with crossOrigin "" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&ignore" with crossOrigin "" should be TAINTED assert_equals: expected "TAINTED" but got "LOAD_ERROR"
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&ignore" with crossOrigin "anonymous" should be LOAD_ERROR
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&ACAOrigin=https://localhost:9443&ignore" with crossOrigin "anonymous" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&ignore" with crossOrigin "use-credentials" should be LOAD_ERROR
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&ACAOrigin=https://localhost:9443&ignore" with crossOrigin "use-credentials" should be LOAD_ERROR
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&ACAOrigin=https://localhost:9443&ACACredentials=true&ignore" with crossOrigin "use-credentials" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&Auth&ignore" with crossOrigin "" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&Auth&ignore" with crossOrigin "" should be TAINTED assert_equals: expected "TAINTED" but got "LOAD_ERROR"
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&Auth&ignore" with crossOrigin "anonymous" should be LOAD_ERROR
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&Auth&ignore" with crossOrigin "use-credentials" should be LOAD_ERROR
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&Auth&ACAOrigin=https://localhost:9443&ignore" with crossOrigin "use-credentials" should be LOAD_ERROR
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&Auth&ACAOrigin=https://localhost:9443&ACACredentials=true&ignore" with crossOrigin "use-credentials" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=same-origin&url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=same-origin&url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "anonymous" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=same-origin&url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "use-credentials" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=same-origin&url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=same-origin&url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "anonymous" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=same-origin&url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "use-credentials" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=no-cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "" should be TAINTED assert_equals: expected "TAINTED" but got "LOAD_ERROR"
PASS url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=no-cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "anonymous" should be LOAD_ERROR
PASS url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=no-cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "use-credentials" should be LOAD_ERROR
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=no-cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "" should be TAINTED assert_equals: expected "TAINTED" but got "LOAD_ERROR"
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=no-cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "anonymous" should be LOAD_ERROR
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=no-cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO" with crossOrigin "use-credentials" should be LOAD_ERROR
PASS url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "" should be LOAD_ERROR
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&credentials=same-origin&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "anonymous" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
PASS url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "use-credentials" should be LOAD_ERROR
FAIL url "https://localhost:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACACredentials%3Dtrue%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "use-credentials" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "" should be LOAD_ERROR
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&credentials=same-origin&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "anonymous" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
PASS url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "use-credentials" should be LOAD_ERROR
FAIL url "https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?VIDEO&mode=cors&url=https%3A%2F%2F127.0.0.1%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FVIDEO%26ACACredentials%3Dtrue%26ACAOrigin%3Dhttps%3A%2F%2Flocalhost%3A9443" with crossOrigin "use-credentials" should be NOT_TAINTED assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
PASS restore global state
